added exceptions and refactored cityController

+++ src/main/java/com/sparta/zmsb/weekfiveteamproject/controllers/CityController.java
@@ -2,6 +2,7 @@ package com.sparta.zmsb.weekfiveteamproject.controllers;
 
 import com.sparta.zmsb.weekfiveteamproject.entities.CityEntity;
 import com.sparta.zmsb.weekfiveteamproject.entities.CountryEntity;
+import com.sparta.zmsb.weekfiveteamproject.exceptions.InvalidInputException;
 import com.sparta.zmsb.weekfiveteamproject.exceptions.ResourceNotFoundException;
 import com.sparta.zmsb.weekfiveteamproject.service.WorldService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -37,97 +38,65 @@ public class CityController {
         this.worldService = worldService;
     }
 
-    @Operation(summary = "Post endpoint", responses = {
-            @ApiResponse(responseCode = "200", description = "Successful post"),
-            @ApiResponse(responseCode = "404", description = "Resource not found")
-    })
+    @Operation(summary = "Post endpoint", responses = {@ApiResponse(responseCode = "200", description = "Successful post"), @ApiResponse(responseCode = "404", description = "Resource not found")})
     @PostMapping
     public ResponseEntity<EntityModel<CityEntity>> createCity(@RequestBody @Valid CityEntity cityEntity, HttpServletRequest request) {
         List<CountryEntity> countries = worldService.allCountries();
 
         countries = countries.stream().filter(c -> c.getCode().equals(cityEntity.getCountryCode().getCode())).toList();
 
-        if (!countries.getFirst().getCode().equals(cityEntity.getCountryCode().getCode()) ) {
-            try {
-                throw new ResourceNotFoundException("Country with code: " + cityEntity.getCountryCode().getCode() + " does not exist");
-            } catch (ResourceNotFoundException e) {
-                throw new RuntimeException(e);
-            }
+        if (countries.isEmpty()) {
+            throw new ResourceNotFoundException("Country with code: " + cityEntity.getCountryCode().getCode() + " does not exist");
         }
 
-        List<EntityModel<CityEntity>> cityEntityModel = Stream.of(worldService.createCity(cityEntity))
-                .map(
-                        city ->
-                        {
-                            List<Link> countryLinks =
-                                    Stream.of(city.getCountryCode().getCode())
-                                            .map(
-                                                    code -> WebMvcLinkBuilder.linkTo(
-                                                            methodOn(CountryController.class).getCountry(city.getCountryCode().getCode())).withRel(city.getCountryCode().getName()))
-                                            .toList();
-                            Link selfLink = WebMvcLinkBuilder.linkTo(
-                                    methodOn(CityController.class).getCity(city.getId())).withSelfRel();
-                            Link relLink = WebMvcLinkBuilder.linkTo(
-                                    methodOn(CityController.class).getAllCities()).withRel("city");
-                            return EntityModel.of(city, selfLink, relLink).add(countryLinks);
-                        })
-                .toList();
+        List<EntityModel<CityEntity>> cityEntityModel = Stream.of(worldService.createCity(cityEntity)).map(city -> {
+            List<Link> countryLinks = Stream.of(city.getCountryCode().getCode()).map(code -> WebMvcLinkBuilder.linkTo(methodOn(CountryController.class).getCountry(city.getCountryCode().getCode())).withRel(city.getCountryCode().getName())).toList();
+            Link selfLink = WebMvcLinkBuilder.linkTo(methodOn(CityController.class).getCity(city.getId())).withSelfRel();
+            Link relLink = WebMvcLinkBuilder.linkTo(methodOn(CityController.class).getAllCities()).withRel("city");
+            return EntityModel.of(city, selfLink, relLink).add(countryLinks);
+        }).toList();
 
         URI location = URI.create(request.getRequestURL().toString() + "/" + cityEntity.getId());
 
         return ResponseEntity.created(location).body(cityEntityModel.getFirst());
     }
 
-    @Operation(summary = "Get endpoint", responses = {
-            @ApiResponse(responseCode = "200", description = "Successful get")
-    })
+    @Operation(summary = "Get endpoint", responses = {@ApiResponse(responseCode = "200", description = "Successful get")})
     @GetMapping
     public CollectionModel<EntityModel<CityEntity>> getAllCities() {
-        List<EntityModel<CityEntity>> cities = worldService.allCities()
-                .stream()
-                .map(
-                        city ->
-                        {
-                            List<Link> countryLinks =
-                                    Stream.of(city.getCountryCode().getCode())
-                                            .map(
-                                                    code -> WebMvcLinkBuilder.linkTo(
-                                                            methodOn(CountryController.class).getCountry(city.getCountryCode().getCode())).withRel(city.getCountryCode().getName()))
-                                            .toList();
-                            Link selfLink = WebMvcLinkBuilder.linkTo(
-                                    methodOn(CityController.class).getCity(city.getId())).withSelfRel();
-                            Link relLink = WebMvcLinkBuilder.linkTo(
-                                    methodOn(CityController.class).getAllCities()).withRel("city");
-                            return EntityModel.of(city, selfLink, relLink).add(countryLinks);
-                        })
-                .toList();
+        List<EntityModel<CityEntity>> cities = worldService.allCities().stream().map(city -> {
+            List<Link> countryLinks = Stream.of(city.getCountryCode().getCode()).map(code -> WebMvcLinkBuilder.linkTo(methodOn(CountryController.class).getCountry(city.getCountryCode().getCode())).withRel(city.getCountryCode().getName())).toList();
+            Link selfLink = WebMvcLinkBuilder.linkTo(methodOn(CityController.class).getCity(city.getId())).withSelfRel();
+            Link relLink = WebMvcLinkBuilder.linkTo(methodOn(CityController.class).getAllCities()).withRel("city");
+            return EntityModel.of(city, selfLink, relLink).add(countryLinks);
+        }).toList();
         return CollectionModel.of(cities, WebMvcLinkBuilder.linkTo(methodOn(CityController.class).getAllCities()).withSelfRel());
     }
 
     @GetMapping("/{id}")
     public CollectionModel<EntityModel<CityEntity>> getCity(@PathVariable @Valid Integer id) {
-        List<EntityModel<CityEntity>> cityEntityModel = Stream.of(worldService.getCityById(id))
-                .map(
-                        city ->
-                        {
-                            List<Link> countryLinks =
-                                    Stream.of(city.getCountryCode().getCode())
-                                            .map(
-                                                    code -> WebMvcLinkBuilder.linkTo(
-                                                            methodOn(CountryController.class).getCountry(city.getCountryCode().getCode())).withRel(city.getCountryCode().getName()))
-                                            .toList();
-                            Link selfLink = WebMvcLinkBuilder.linkTo(
-                                    methodOn(CityController.class).getCity(city.getId())).withSelfRel();
-                            Link relLink = WebMvcLinkBuilder.linkTo(
-                                    methodOn(CityController.class).getAllCities()).withRel("city");
-                            return EntityModel.of(city, selfLink, relLink).add(countryLinks);
-                        })
-                .toList();
+
+        CityEntity cityEntity = worldService.getCityById(id);
+        if (cityEntity == null) {
+            throw new ResourceNotFoundException("City with ID: " + id + " not found");
+        }
+
+        List<EntityModel<CityEntity>> cityEntityModel = Stream.of(worldService.getCityById(id)).map(city -> {
+            List<Link> countryLinks = Stream.of(city.getCountryCode().getCode()).map(code -> WebMvcLinkBuilder.linkTo(methodOn(CountryController.class).getCountry(city.getCountryCode().getCode())).withRel(city.getCountryCode().getName())).toList();
+            Link selfLink = WebMvcLinkBuilder.linkTo(methodOn(CityController.class).getCity(city.getId())).withSelfRel();
+            Link relLink = WebMvcLinkBuilder.linkTo(methodOn(CityController.class).getAllCities()).withRel("city");
+            return EntityModel.of(city, selfLink, relLink).add(countryLinks);
+        }).toList();
         return CollectionModel.of(cityEntityModel, WebMvcLinkBuilder.linkTo(methodOn(CityController.class).getAllCities()).withSelfRel());
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<EntityModel<CityEntity>> updateCity(@PathVariable @Valid Integer id, @RequestBody @Valid CityEntity cityEntity) {
+
+        if (!id.equals(cityEntity.getId())) {
+            throw new InvalidInputException("City ID in path: " + id + " does not match ID in payload: " + cityEntity.getId());
+        }
+
         List<CityEntity> cities = worldService.allCities();
         List<CountryEntity> countries = worldService.allCountries();
 
@@ -136,8 +105,8 @@ public class CityController {
 
         if (!id.equals(cityEntity.getId())) {
             try {
-                throw new InvalidEndpointException("City ID: " + cityEntity.getId() + " does not match the endpoint ID: " + id);
-            } catch (InvalidEndpointException e) {
+                throw new ResourceNotFoundException("City ID: " + cityEntity.getId() + " does not match the endpoint ID: " + id);
+            } catch (ResourceNotFoundException e) {
                 throw new RuntimeException(e);
             }
         } else if (cities.isEmpty()) {
@@ -146,7 +115,7 @@ public class CityController {
             } catch (ResourceNotFoundException e) {
                 throw new RuntimeException(e);
             }
-        } else if (!countries.getFirst().getCode().equals(cityEntity.getCountryCode().getCode()) ) {
+        } else if (!countries.getFirst().getCode().equals(cityEntity.getCountryCode().getCode())) {
             try {
                 throw new ResourceNotFoundException("Country with code: " + cityEntity.getCountryCode().getCode() + " does not exist");
             } catch (ResourceNotFoundException e) {
@@ -160,14 +129,13 @@ public class CityController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<CollectionModel<EntityModel<CityEntity>>> deleteCity(@PathVariable @Valid Integer id) {
-        if (worldService.getCityById(id) == null) {
-            try {
-                throw new InvalidEndpointException("Endpoint: " + id + " does not correspond to a City");
-            } catch (InvalidEndpointException e) {
-                throw new RuntimeException(e);
-            }
+        CityEntity city = worldService.getCityById(id);
+        if (city == null) {
+            throw new ResourceNotFoundException("City with ID: " + id + " not found");
         }
+
         worldService.deleteCity(id);
+
         return ResponseEntity.noContent().build();
     }
